Random component consolidation
