--- src/components/DurationInput/index.ts
@@ -1,1 +1,0 @@
-export * from './DurationInput'
