+++ src/components/IconedImage.tsx
