+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/ProjectSettingsContent.tsx
@@ -8,7 +8,7 @@ import { useMemo } from 'react'
 import { ArchiveProjectSettingsPage } from './pages/ArchiveProjectSettingsPage'
 import { GovernanceSettingsPage } from './pages/GovernanceSettingsPage'
 import { PayoutsSettingsPage } from './pages/PayoutsSettingsPage'
-import { ProjectDetailsSettingsPage } from './pages/ProjectDetailsSettingsPage'
+import { ProjectDetailsSettingsPage } from './pages/ProjectDetailsSettingsPage/ProjectDetailsSettingsPage'
 import { ProjectHandleSettingsPage } from './pages/ProjectHandleSettingsPage'
 import { ProjectUpgradesPage } from './pages/ProjectUpgradesPage'
 import { ReconfigureFundingCycleSettingsPage } from './pages/ReconfigureFundingCycleSettingsPage'
